% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{TraitRef}
\alias{TraitRef}
\title{Reference about trait}
\format{
A data frame with 144 rows and 8 variables:
\describe{
  \item{Number}{Reference number}
  \item{Title}{Reference titles}
  \item{Author}{Author}
  \item{Year}{Year}
  \item{Publisher}{Publisher}
  \item{Published City}{Published City}
  \item{ISBN}{ISBN}
  \item{DOI}{DOI}
}
}
\usage{
TraitRef
}
\description{
A dataset containing the measured height of four species of trees
}
\keyword{datasets}
